Fix regression in bases Split merging into separate class
