+++ src/Analysis/Ast/Impl/Specializations/Specialized.cs
@@ -19,7 +19,7 @@ namespace Microsoft.Python.Analysis.Specializations {
     internal static class Specialized {
         public static IPythonPropertyType Property(string name, IPythonModule declaringModule, IPythonType declaringType, IMember returnValue) {
             var location = new Location(declaringModule);
-            var prop = new PythonPropertyType(name, location, declaringType, false);
+            var prop = new PythonPropertyType(name, location, string.Empty, declaringType, false);
             var o = new PythonFunctionOverload(prop, location);
             o.AddReturnValue(returnValue);
             prop.AddOverload(o);
